+++ src/LaravelConfigServiceProvider.php
@@ -25,18 +25,6 @@ class LaravelConfigServiceProvider extends ServiceProvider
             $this->publishes([
                 __DIR__.'/../database/factories/ConfigFactory.php' => database_path('factories/ConfigFactory.php'),
             ], 'laravel-config-factories');
-
-            if (! class_exists('CreateLaravelConfigTable')) {
-                $this->publishes([
-                    __DIR__.'/../database/migrations/create_laravel_config_table.php.stub' => database_path('migrations/'.date('Y_m_d_His').'_create_laravel_config_table.php'),
-                ], 'laravel-config');
-            }
-
-            if (! class_exists('AddTagsColumnToConfigTable')) {
-                $this->publishes([
-                    __DIR__.'/../database/migrations/add_tags_column_to_config_table.php.stub' => database_path('migrations/'.date('Y_m_d_His').'_add_tags_column_to_laravel_config_table.php'),
-                ], 'laravel-config');
-            }
         }
     }
 
@@ -46,7 +34,7 @@ class LaravelConfigServiceProvider extends ServiceProvider
     public function register(): void
     {
         // Automatically apply the package configuration
-        // $this->mergeConfigFrom(__DIR__.'/../config/config.php', 'laravel-config');
+        $this->mergeConfigFrom(__DIR__.'/../config/config.php', 'laravel-config');
 
         // Register the main class to use with the facade
         $this->app->singleton('laravel-config', function () {