+++ tests/TestCase.php
@@ -13,7 +13,12 @@ class TestCase extends \Orchestra\Testbench\TestCase
 
         $this->withFactories(__DIR__.'/../database/factories');
 
-        $this->setUpDatabase($this->app);
+        Schema::dropAllTables();
+
+        $this->artisan('migrate', [
+            '--database' => 'mysql',
+            '--realpath' => realpath(__DIR__.'/../database/migrations'),
+        ]);
     }
 
     /**
@@ -26,16 +31,5 @@ class TestCase extends \Orchestra\Testbench\TestCase
         return [
             LaravelConfigServiceProvider::class,
         ];
-    }
-
-    protected function setUpDatabase($app)
-    {
-        Schema::dropAllTables();
-
-        include_once __DIR__.'/../database/migrations/create_laravel_config_table.php.stub';
-        include_once __DIR__.'/../database/migrations/add_tags_column_to_config_table.php.stub';
-
-        (new \CreateLaravelConfigTable)->up();
-        (new \AddTagsColumnToConfigTable)->up();
     }
 }